# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GitHut - Git Server API",
    "description": "A self-hosted Git server API that provides repository management, authentication, SSH key management, and Git Smart HTTP protocol support.",
    "version": "1.0.0",
    "contact": {
      "name": "Bravo68Web"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user",
        "description": "Returns the currently authenticated user's information",
        "operationId": "getCurrentUser",
        "security": [{ "bearerAuth": [] }, { "basicAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ssh-keys": {
      "get": {
        "tags": ["SSH Keys"],
        "summary": "List SSH keys",
        "description": "Returns all SSH keys for the authenticated user",
        "operationId": "listSSHKeys",
        "security": [{ "bearerAuth": [] }, { "basicAuth": [] }],
        "responses": {
          "200": {
            "description": "List of SSH keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSSHKeysResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["SSH Keys"],
        "summary": "Add SSH key",
        "description": "Adds a new SSH public key for the authenticated user",
        "operationId": "addSSHKey",
        "security": [{ "bearerAuth": [] }, { "basicAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSSHKeyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "SSH key added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddSSHKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid SSH key format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "SSH key already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ssh-keys/{id}": {
      "get": {
        "tags": ["SSH Keys"],
        "summary": "Get SSH key",
        "description": "Returns a specific SSH key by ID",
        "operationId": "getSSHKey",
        "security": [{ "bearerAuth": [] }, { "basicAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "SSH key ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSH key information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SSHKeyInfo"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "SSH key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["SSH Keys"],
        "summary": "Delete SSH key",
        "description": "Deletes an SSH key by ID",
        "operationId": "deleteSSHKey",
        "security": [{ "bearerAuth": [] }, { "basicAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "SSH key ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSH key deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "SSH key deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "SSH key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token authentication"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Basic authentication with username and password"
      }
    },
...
    "schemas": {
...
      "UserInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "username": {
            "type": "string",
            "example": "johndoe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "is_admin": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "AddSSHKeyRequest": {
        "type": "object",
        "required": ["title", "key"],
        "properties": {
          "title": {
            "type": "string",
            "description": "A descriptive name for the SSH key",
            "example": "My Laptop"
          },
          "key": {
            "type": "string",
            "description": "The public SSH key in OpenSSH format",
            "example": "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIExample... user@laptop"
          }
        }
      },
      "AddSSHKeyResponse": {
        "type": "object",
        "properties": {
          "key": {
            "$ref": "#/components/schemas/SSHKeyInfo"
          },
          "message": {
            "type": "string",
            "example": "SSH key added successfully"
          }
        }
      },
      "SSHKeyInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "example": "My Laptop"
          },
          "fingerprint": {
            "type": "string",
            "description": "SHA256 fingerprint of the public key",
            "example": "SHA256:abc123def456ghi789jkl012mno345pqr678stu901v"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
      "ListSSHKeysResponse": {
        "type": "object",
        "properties": {
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SSHKeyInfo"
            }
          },
          "total": {
            "type": "integer",
            "example": 2
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "not_found"
          },
          "message": {
            "type": "string",
            "example": "Repository not found"
          },
          "details": {
            "type": "string",
            "example": "Additional error details"
          }
        }
      }
    }
  }
}